refactor: Factory moved in factory folder
